+++ pkg/sql/colexec/lockop/types.go
@@ -51,7 +51,7 @@ type LockOptions struct {
 	maxCountPerLock          int
 	mode                     lock.LockMode
 	sharding                 lock.Sharding
-	group                    uint32
+	group                    int64
 	lockTable                bool
 	changeDef                bool
 	parker                   *types.Packer
