+++ pkg/vm/engine/memoryengine/access_info.go
@@ -21,7 +21,7 @@ import (
 )
 
 type AccessInfo struct {
-	AccountID uint32
+	AccountID int64
 	UserID    uint32
 	RoleID    uint32
 }
